tests/trip: update for current results from TeXLive XeTeX
The new reference files are generated by a build of TeXLive XeTeX, version
3.14159265-2.6-0.99996. There are differences from what's stored in the
official repository, but none of them seem problematic (as you'd hope). Many
of the differences come because of wrapping at 80 rather than 72 columns.
Others come from the fact the XeTeX treats a charcode of 256 as being an
actual Unicode character, while TeX often treats it as in invalid character
that should not be printed.

The files were generated with this command:

  xetex -fmt=trip -no-pdf trip >trip.fot