Fix a few subtle problems exposed in the trip test.
When we put underscores back in the variable names, various elements of the
web2c pipeline broke because they hardcoded checks for various TeX variable
names. I thought I got most of them in fixwrites, but it turns out there was a
subtle and pernicious hack in the web2c parser that dealt with exiting from a
couple of functions. All of this stuff makes me scared that there are more
problems lurking, but hey, we pass the trip test now.