Disable the optional "STAT" code.
According to the TeXBook, this collects memory usage statistics and
information for \tracingparagraphs and \tracingpages. I don't care about that
stuff.

Disabling STAT causes the output to change fairly significantly, though,
meaning that we have to update the TRIP and eTRIP files. It's not ideal to
lose the diagnostic information, but I think the tests will still be very
effective for identifying when we break things. And it's just never going to
be possible for us to pass a pristine TRIP test since it's defined in such an
implementation-specific way.