Finally stop hardcoding the path to the bundle!
Now that we have command-line argument handling in Rust, it is feasible to
make the bundle path a runtime option. We've also gained the ability to
operate without a bundle, just using the local filesystem to find files.
That's sufficient for the TRIP and eTRIP tests.

The I/O layer totally needs a lot of cleaning up to be more flexible, but this
gets rid of a big gross hardcoded absolute path that meant that no one else
was going to be able to run anything successfully.

To be resolved: still not sure how to make it so that the driver can
automatically find a sensible default bundle. Probably have it search for
something sensible in a ~/.share/ or something.