Route all writes to stdout through the Rust I/O layer.
A lot less to this than anticipated. If I tweak the Rust code to simply ignore
writes to stdout, TeX's chatter disappears as expected.

The OutputItem enum used to handle the stdout output method is something of a
hack. I expect it to disappear once the Rust output handling is formalized.