Switch all image I/O over to Rust-y I/O.
This is a lot of code churn and something is likely busted. But at the time
being I don't have nice test cases to debug this. I get the same results on my
xenia test as before the changes, so at least the PDF probing code isn't
totally bonkers.

Not many FILE*'s left in the C code ...