tests/trip.rs: merge the eTRIP test in here
This exposes the problem that the Rust test framework wants to run everything
in simultaneous threads, which is really bad news for us. This can be worked
around with the RUST_TEST_THREADS environment variable or the --test-threads
option, but I want everything to Just Work automatically. So, I add a fun
global static mutex to have the engines not step on each other's toes.