tectonic/XeTeXFontInst.cpp: avoid common kpsezip complaints
The AFM-loading code here was called commonly, leading to lots of complaints
because it hadn't been ported to the new Rust I/O layer. I've attempted to
port it over, although I suspect that this code path isn't ever actually being
executed so who knows. This is just the very beginning of us grappling with
how to deal with system font files in the context of our efforts to make the
engine as self-contained as it can be.