Make io::IOStack take references, not owned boxes.
Since this was taking owned boxes of trait objects, we lost the ability to
know the specific types of the elements of the IO stack, making it impossible
to, e.g., access the contents of a MemoryIO after running the engine ... at
least without hacks to duplicate its "files" field.

I was worried that this wouldn't be possible because of the interaction
between the needed lifetime parameter and the global Engine pointer used to
handle the C callbacks, but it turns out our fakery gets the job done.