Woo, shut up TeX.
First we modify Engine::process() to return the detailed success level of the
processing run. Then, the driver program looks at this and decides what to do
based on the outcome. Output files are buffered in memory and not actually
written to disk unless the engine is successful. The log file is thrown away
by default. The stdout chatter is suppressed by default, but printed if the
engine errors out. Finally, halt_on_error_p is set to true so that things that
normal people would consider to be errors are treated as such, stopping the
engine immediately.