tectonic/XeTeXFontMgr_FC.cpp: attempt to solve the TRIP test segfault
The TRIP test, which runs two engines in separate but mutually-excluded
threads, sometimes dies with a segfault. Inspection of a coredump suggests
that the problem is in the cleanup in XeTeXFontMgr_FC::terminate(). That
function doesn't clear the pointers that it zeros out, which seems like the
kind of thing that could cause problems. I haven't reproduced the problem with
this patch, though that's only so meaningful given its intermittency. Let's
see if it helps though.