src/itarbundle.rs: new bundle implementation
This uses an "indexed tar" file as the storage backend. The backing file
actually doesn't have to be a tar at all; it just needs to have each bundled
file somewhere in the stream that we can just seek to it and read it.

This was done because the Zip file turns out not to be as good for Web remote
access as I envisioned; to get the information for each file you need to read
its local Zip header, which means you need to seek through the entire Zip file
when you initially open it, since standard practice is to gather information
about all of the files in the Zip when it's opened. The "index file" approach
collects all of this header information in one place.

Note that the backing tar file can't be compressed since we need to be able to
seek around in it. It might be better to use a Zip file as the backing data
store since in Zips the individual items are compressed separately, so you
could seek to an item and decompress it on the fly. But that would be a more
complicated undertaking since at that point you're basically writing a Zip
reader.