src/itarbundle.rs: restructure the backing HTTP implementation
Just request files all at once, instead of trying to be clever with caching
and chunking. This is a lot faster.

I was also hoping that we could just hand off HTTP response streams to the
callers, but I forgot that we currently want our input streams to be seekable,
which wouldn't work. And there are lifetime issues associated with returning
the response streams.