Wire up the other kpathsea format types into the Rust code.
Might as well just take care of them. I thought that the "pict" format should
have more extensions like "jpg" associated with it, but I can't figure out
if/how that's done. It might be that that sort of extension-guessing happens
at the TeX level, not in the C code.