Remove #if LIBDPX code
This sounds very tempting, but it introduces a few elements that I don't quite
understand, and the XeTeX results that I want to emulate are definitely based
on the non-LIBDPX code paths. It only does a few simple library-ification
things that I can easily reproduce if needed.

One of the #defines leads me to believe that the LIBDPX changes are associated
with "ApTeX". I do not know what this is and I'm on a plane so I can't google
it.