Start making the engine interact with IO more sensibly.
I think I figured out the main problem: the Engine shouldn't own the IO layer.
It should just borrow it while the processing is running.

The solution here is lame and bad; I think I know how to do better. But the
code in this commit is enough in the right direction that I think it's worth
committing.