Switch the IOStack type to take a slice, not a Vec.
No particular need for us to take ownership in the better model that we're
using where the Engine merely borrows its IO backend. The type that we're
using for the slice doesn't feel right -- &'a mut [&'a mut IOProvider] -- but
that's the only one I can get to compile right now.