tectonic/jpegimage.c: fix this
I'm not sure why I got rid of the logic that was preventing Exif from
overriding JFIF info, but it was causing JPEGs to come out mis-sized since the
xdvipdfmx equivalent code still had it.

This is caused by the redundant image-reading code on the xdvipdfmx and xetex
sides, which is silly. I will work on making that go away.