Implement the full local cache scheme.
First, we have a configuration file with a default bundle URL programmed in.
Second, we add a layer that caches the SHA1 digest of the bundle found at the
URL, so that we can operate without requiring a network connection (at the
risk of not realizing that the backing bundle has been updated).