Pass the StatusBackend through to IoProvider::input_open_name().
And it turns out that traits can't be turned into trait objects if they have
generic parameters. We reallly want IoProviders to be usable as trait objects
so that we can put them in generic IoStacks, so we can't have the IoProvider
methods be generic over the StatusBackend type. Therefore we need to pass the
StatusBackend as a trait object. Therefore we don't need the engine
ExecutionState object to be generic of StatusBackend after all.