bibtex: fix error in C transcription
Ah, there was a non-local goto in the original code that can't translate into C.
It's related to the "lab31" variable that I removed early on, although that
variable definitely didn't do anything. I wouldn't be surprised if there was
an underscorey-type issue with manual fixup in the C translation that I missed.
Anyway, not too hard to obtain the desired effect.