Fix no_pdf_output = 1 in the TeX engine.
We have had this fixed for a while. Tidies up various bits of code. It does
turn out, though, that Synctex depends on this variable in a way that affects
its output. Synctex is very much a back-burner thing so I'll just hardcode the
value now and I can go back a look at this commit later if needed.