Convert OutputHandle into a thicker wrapper struct
In particular we add on a SHA256 digest computation as we write the output.
This is another element of being able to tell when TeX needs to be rerun. If a
file has a read-then-write access pattern, and what was written is different
than what was read, the processing hasn't converged.