cli: avoid circular deps in the output Makefile rules
It was possible for the Makefile code to emit circular rules, if e.g. the .aux
file was rewritten by the engine and the modified .aux file was actually saved
to disk. We now identify and avoid that case, while we don't complain if the
"modified" file isn't actually written to disk, which is often going to be the
case with .aux files and the like.

(The old FileSummary struct was the natural place to keep the needed
housekeeping information, which was what motivated the refactoring into the
IoEventBackend design.)