cli: add a "bibtex_first" pass mode
This is like the default mode, but it unconditionally runs bibtex first, then
applies the standard rerun logic. This is just so that I can implement an
efficient Makefile-like pipeline for compiling TeX documents where I generate
the .bib programmatically:

1) Run tectonic --pass=tex --outfmt=aux to write the aux file
2) Run my program to process the .aux file and write the .bib
3) Run tectonic --pass=bibtex_first to finish the compilation.

All this feature does is save one potentially-unneeded TeX invocation at the
beginning of Step 3, but it's nice to be efficient.