cli: make the "aux" output format really only output .aux files
Blah. And I thought I was being so clever before. And I was! But, Ninja is not
clever: it does not appear that it can handle a Makefile deps file that writes
out rules for more than one file. If you emit rules of the form `file1 file: a
b c`, it errors out, but if you emit two *separate* rules, it erroneously
merges everything after the first. This is now Ninja bug #1258 on GitHub.

The problem was that in my use case, the --outfmt=aux mode ended up emitting
both a .aux and a .out file. Ninja needs Tectonic to choose the .aux file as
the one "real" output, but Tectonic has no way of knowing that. So with the
previous behavior, my Ninja build system use case wasn't going to work.

The easiest solution is just to have the "aux" output format really only emit
.aux files. Comparable uses can be covered by the `-k` option so I'm pretty OK
with this different approach.