Fix output of files that shrink.
Duhhhh. I ran into a problem with corrupt XDV output files. It turns out that
the memory-backed I/O provider had a problem where it would re-use buffers for
*output* files as well as *input* files. This meant that if you created a file
in one pass of the TeX engine, then wrote out a shorter version of it in a
subsequent pass, you'd get leftover junk at the end. Silly, but easy to fix:
truncate when opening for output.

I'm actually adding a test case for this bug! Look at me, all responsible.