Start generating the "string pool" resources ourselves.
First of all, this will make it more sane to add new string-pool constants,
which I need to do to stub out a no-op \pdfoutput primitive to get Matt's
example files working.

Second, I'm prototyping a system that will allow the use of symbolic constants
for the string-pool constants, which will make the source more readable. Instead
of

      65538L /*"buffer size" */

we'll be able to write

      S(buffer_size)

Third, this should make it easier to eventually transition away from the
string pool model into something a bit saner.