Permanently enable e-TeX.
We're just going to unilaterally declare it to be a thing, although we have a
hack to work with TeXLive format files that may not have e-TeX enabled
(namely, trip.fmt). This alters the TRIP/eTRIP output a little bit but only
the cosmetics; and we've already gone down the path of changing those.