Demonstrate INITEX mode in the (e-)TRIP tests
Now that we can INITEX, we can generate the formats used for the TRIP tests
on-the-fly. For some reason this causes the tests to take *far* longer to run,
but they pass! Nice.

The TRIP test output changes slightly. I'm 99% sure this is because we've made
it so that eTeX mode is always enabled. The impact is minimal and its origin
makes sense (I think ...) so I'm OK with that.