tectonic: go back to making a copy of the eqtb pointer
I am concerned that the TRIP tests are suddenly taking so long to run, and
it's possible that the eqtb pointer caching really does make a difference in
performance. So, just in case, let's go back to duplicating it locally in each
function. The tests still take a very long time to run, unfortunately, but
builds in release mode seem quite fast.