tectonic/XeTeX_ext.c: stop including <png.h>
It's not used here, and it caused problems for GitHub user @newpavlov because
the version of libpng on Ubuntu 14.04 complains if you include <setjmp.h>
before it, and Freetype includes setjmp, and Freetype came before libpng in
that file. But anyway: it's not needed, and getting rid of it gets the project
to build on a 14.04 Docker container for me.