Add a new integer variable: \pdfoutput
At long last, I think we have all of the pieces in place to add this new
variable. Doing so means that we have to split off and start using a new
format file version, which is scary, but I think we're ready. The actual code
itself should be straightforward, though.

It would be cool, though, for the ASSIGN_INT case of prefixed_command() to
gain some logic to identify when a compat-only variable is assigned and issue
a warning, one-time only, that the variable does nothing. I'll save that for
later, though.