Save formats in a special munged path and generate formats on-the-fly.
Took me a long time to get this working but I think this design should work
pretty well. We have the LocalCache IoProvider intercept reads of formats
and add infrastructure so that the CLI driver can tell the LocalCache to
handle the saving of format files directly.

I was hoping to be able to make it so that the output layer could know that
the TeX engine was trying to write a format and save it directly in the
LocalCache format cache directory, but I couldn't get it to work cleanly since
you still need a full IoStack to save the log file, potentially show stdout
output, and so on.