Lock the manifest file when using the local filesystem cache layer.
If two Tectonic processes were to download files simultaneously, they would
append new lines to the manifest file willy-nilly, resulting in interleaved
lines that corrupted the file. We now use file locking via the `fs2` crate to
make sure that these accesses are appropriately serialized.

The results are still somewhat gross: two processes running simultaneously
will not pay attention to each other's work, so they'll download redundant
files and add redundant lines to the manifest. That's not cool, but we can
live with it.

Closes #7.