Remove `emit_tex_constants` sub-crate.
It turns out that to publish things on crates.io, these helper crates must be
individually published, which would really be a bit silly here. So let's just
fold the code into the build.rs.

Part of this is that the support crate ended up involving a lot less code than
I originally envisioned.