tectonic: add a \pdfmdfivesum primitive
Encountered in the wild: a problem with code that seems to think we're pdftex
(perhaps because of the \pdfoutput keyword?) and therefore ends up with a
busted definition of a \pdf@filemdfivesum macro.

I suspect that this is an example of Khaled Hosny's warning that maybe
defining \pdfoutput is a bad idea because things will start thinking that
we're pdftex. But the hope is that it won't be too much trouble to fake
things, so we plunge on ahead.

Since this adds a new primitive and string, this requires a format version
bump. I have not done that here since there are other changes I want to make,
so we are:

[temporarily-broken]