tectonic/xetexini.c: fix format loading
Sigh, when I removed the eTeX detection bit from the format file, I didn't
update a piece of code that set the variable `format_written_with_etex`.
A little annoyed that I didn't get to see a compiler warning about using
an unitialized variable, but so be it.

Also, for some reason this problem only manifested in release builds, not
debug builds, which is why it hung around for a little while.