Remove Poppler dependency (!)
No lack of love for Poppler, but xdvipdfmx already includes code for reading
PDF files and extracting their page bounds, and it turns out that was the only
reason that XeTeX depended on Poppler.

It was not completely trivial to merge XeTeX's code to get the PDF bounding
box with the "example" in xdvipdfmx's `xbb.c`, but I think it's all good here.

Poppler was our heaviest-weight dependency so this is great!