Use a structure of function pointers to express the C/C++ => Rust API.
Instead of defining `ttstub_*` symbols in Rust that are exported to the C/C++
code, we now populate a structure of callbacks and pass that the the C/C++
entry points.

A key advantage is that we can now maintain type information: we don't have to
hardcode that the IoProvider used by the ExecutionState is the IoStack type in
particular, for instance. I'd say that the Rust code is generally a lot more
pleasant.

Now, the global variable that expresses the single C<=>Rust connection is
defined in the C code, rather than on the Rust side of things.