Support amsrefs
This commit changes the condition that decides whether bibtex is run to
finding `\bibdata` in the aux file (was `\citation` and `\bibcite`).
Since bibtex exits with an error if it does not find `\bibdata` anyway,
this avoids running bibtex unecessarily when for instance amsrefs is
used to process the bibliography.