dist/deploy_key.enc: Add an encrypted SSH key for deployment via Travis
This uses Travis' support for encrypted files to add a passwordless RSA SSH
key that can be used to perform various actions when new verisons are
released.

This is intended to support automatic updating of the Arch Linux repository. I
thought about making the filename of the key specific to Arch (i.e., so that
we'd make new keys for each deployment scenario), but it seems more reasonable
and no less secure to just use one key. If a key gets compromised, you'd have
to act as if they all were compromised anyway, I think.