Make a std::os::unix dependency cross-platform
This patch comes with even fewer guarantees than the MIT provides you.
I expect it works, but wchar shenanigans could easily end up happening,
except that none of this software seems to be *designed* to work well on
Windows (it’s all merely ported to Windows—if you’re lucky).