Cargo.toml: attempt to fix us up on docs.rs
A new feature was just merged into docs.rs
(https://github.com/onur/docs.rs/pull/73) which will hopefully make it
possible for Tectonic to build on their system, which would be great.

As far as I can tell, we won't be able to check if this is working until the
next release is made, since docs.rs pulls releases from crates.io. But might
as well get it in now.