tectonic: remove do_early_args() and simplify dvipdfmx_main
No need to use argc & argv[] and its parsing, use explicit arguments
instead. Thus, we do not call do_early_args and do_args anymore and we
can remove do_early_args.

The for loop about the `verbose` variable is left and some bool
variables are left in case it used later. It could also be removed
easily if needed.