tectonic/dpx-dvi: remove handling of config special
Such special would parse the string provided as a command line argument
and would allow to change some parameters of dvipdfmx.

Such feature is not really useful since if a user wants to modify some
parameters we should provide command line parameters to tectonic to do
so directly.

This feature also impedes our ability to simplify the code and remove
global variables.

Removing this feature is equivalent to the parsing of the
dvipdfmx config file which was removed.