build.rs: fix build on Rust 1.18
It seems that in Rust 1.18 (and current 1.19 nightly for what that's worth),
the order in which dependent OS libraries are emitted by build.rs matters more
than it used to. Strangely, regular builds would work fine, but the build of
`tectonic.rlib` for `cargo test` would get a link error because our
`libtectonic_cpp.a` needed a link to `libz`, but the relevant `-lz` flag was
only placed in the command line *before* the command to link in
`libtectonic_cpp.a`, which of course doesn't get the job done in the
order-dependent way that `ld` works.

So, we need to emit the relevant `-lz` flags *after* linking in our own
libraries; but we can't get the correct C include paths for our libraries
without doing the pkg-config call, which is what emits the Cargo metadata for
the OS library links.

The solution is simple: just run pkg-config twice. The `pkg_config` crate has
an option to not print the cargo metadata so we can even suppress its emission
on the first invocation, although I don't think it would be a problem to print
everything twice.

It would be nice if the `pkg_config::Library` object had a method to print the
relevant metadata so that we didn't have to invoke pkg-config twice. But,
really, who cares?