src/errors.rs: implement a lame comparison function
For testing, it is handy to be able to compare errors and results, but
Rust error types don't generically implement PartialEq. Hack around
this with our own DefinitelySame trait that lets us get the right effect,
where false is returned for two objects that we don't *know* to be equal.

In so doing I think I've run into an issue with Rust's trait coherence. I want
to implement DefinitelySame for anything that's PartialEq, but if I do so Rust
thinks that the impl conflicts with the one for std::Result, even though
std::Result is only PartialEq if its subtypes are PartialEq, as I understand
it. For now we just hack around it.