Rework how the "primary input" is handled.
The "primary input" is (as of this commit) the initial stream that the TeX
engine starts parsing. To date, the engine was given a file name to parse, and
it tried to open that file using the I/O subsystem and went from there.

This was dissatisfying for two reasons. First, it meant that if you wanted to
synthesize a primary input for the TeX engine (e.g., stdin or a text string in
memory), you had to make up a "filename" for it and wire that in to the I/O
system. This felt fragile and inelegant.

Second, even if the primary input file was a genuine file on disk, we had to
make sure that it would parse properly if it contained special characters --
the previous code expanded tokens as it parsed the filename so all sorts of
crazy things could happen. And, although this is definitely an edge case, we
wouldn't be able to handle non-Unicode Unix filenames.

This system adds a new I/O API requesting to open the primary input stream in
particular -- a mirror of `output_open_stdout`, in a certain sense. The Rust/C
bridge has then been updated to allow the Rust code to start up I/O on the
primary input using this API and circumventing the standard file-opening
shenanigans, which basically pretend that the user has just typed "\input $1".

As an example of how this makes life better, I'll soon add better support for
TeX processing of a document delivered on stdin. I also believe that this
infrastructure will improve our ability to handle relative paths in documents.