Teach the CLI how to process standard input.
If you specify "-" as the input filename, the CLI will buffer standard input
and process it, including looping through it multiple times if needed. The
outputs will be named "texput" and land in the current directory.

The previous commit was quite complicated, but this one isn't so bad, is it?