Be more careful about paths in the CLI.
Now that the way that we pass the input file's path to TeX has been
straightened out a bit, we can have the CLI be smarter about input paths.

First of all, we adopt the discussed schema that the filesystem location of
the main input file is treated as a virtual "CWD". If you're processing
"foo/bar.tex" and it does "\input other.tex", we will attempt to open the file
"foo/other.tex".

Second, we are more careful about the paths that we send to tex. In the
example above it gets told that the input file name is "bar.tex", and if there
is any Unicode zaniness it gets ironed out. *That* path is what TeX uses to
determine the names of the output files it creates, so in the CLI we adapt
them appropriately.

Finally, when writing output files we now also anchor them to the input. So in
the above example, the final output will be "foo/bar.pdf".

I think it all works.

Closes #31.