tectonic/xetex0.c: fix a signed overflow warning that shows up in --release
The original code gave me this warning:

```
tectonic/xetex0.c: In function ‘conditional’:
tectonic/xetex0.c:10260:12: warning: assuming signed overflow does not occur
  when assuming that (X + c) < X is always false [-Wstrict-overflow]
```

The origin of the warning is a bit difficult for me to understand. But, in the
surrounding code, first `m` is assigned the value of `first`, and then `m` is
possibly incremented multiple times (or maybe once, or maybe never). So, `m >=
first`. Therefore, the clauses in the "if" statement can be reordered in such
a way as to remove the inequalities, which makes the warning go away.