src/engines/mod.rs: don't issue a warning when trying to MD5 a nonexistent file
The standard LaTeX rerun file check modules trigger this warning every build,
which gets annoying. So in the name of UX, don't print anything out. If we had
it, you could imagine issuing a message at a "debug" level of output that
would only be triggered if the user asked for really detailed diagnostics.