build.rs: add flag_if_supported utility function
It allows us to check whether a flag is supported by the compiler
before adding it to the list of flags.

Such feature is useful to support multiple compilers (and multiple
versions of these compilers) for the users compiling the project while
giving the ability for the developpers to add warnings flags.

No equivalent function exists in gcc-rs right now so we implement our own. It
might be interesting to see in the future if a similar feature could be
added in gcc-rs directly.