tectonic: define ZLIB_CONST to avoid casting away const qualifier
zlib requires to define ZLIB_CONST in order for it to use const
qualifiers in its headers. We define it so that when we give it a const
pointer, it is not cast away as non-const.

In addition, this commit removes two useless casts (implicit cast from
void pointers to any other pointer).