tectonic: fix remaining -Wcast-qual warnings
It does not seem there is a need to return a char * from strstartswith
at the moment, so let's only return a const char *.

This commit also removes a useless cast by directly assigning the
variable as const from the beginning, since it is a const.

These fixes allow us to enable -Wcast-qual globally.