tectonic: add approriate attribute to global variables declarations
The global variables from xetexd.h are used across multiple compilation
unit. In such case (when it is not a static variable), one has to declare
the variables with the extern attribute (e.g., in a header file) and to
define them *once* in a C file (without the extern attribute).

Otherwise, multiple definitions of the same global variable is an
undefined behavior.

Other cases are static variables that missed the static attribute.

These fixes allows us to enable -Wmissing-variable-declarations