Start real work on HTML output!
The tentative first step is to wire up a mode that alters the XDV output in a
way that will eventually provide what we need to go a really good HTML
translation. For now, we just alter its identifying byte and file extension.

Format file version bump needed since the string pool changed.