Add the staging repo as a submodule.
This lets us track the precise version of the reference sources on which we
are based at any given time. See discussion in the README.

Fortunately, one does not need to check out or build the reference sources to
build Tectonic, which is good because Git submodules are super confusing.