Synchronize us with staging c35452c4.
This catches us up to TeXLive SVN 45838, or 2017 Nov 17. Many changes since it
has been a long time since I synchronized! I have tried to do everything
correctly but there was a lot of churn.

Somewhat gratifying though is that we seem to have fixed a problem that snuck
into the TRIP test. Comparing to the `trip.log` maintained by TeXLive, the new
version of the file is now *closer* to the original one.

There are a bunch of changes to Synctex to support PDF forms that I tried to
implement, but the code is essentially untested so it probably does not yet
work. (By my reading, if you try to enable PDF forms in the TeXLive version,
it won't even compile.)