Fix a subtle problem in PDF output testing.
It turns out that `xdvipdfmx` will, by default, munge TeX font names with a
"unique tag" in the PDFs that it creates. I don't know why. It makes this tag
using `rand()`, seeding the random number generator with the current time or
the `SOURCE_DATE_EPOCH` environment variable. We already set
`SOURCE_DATE_EPOCH` so the random number generator was getting seeded the same
way, but we cannot expect the generator to give consistent results across
platforms.

Therefore we add a "deterministic" tag mode where the tags are simple integers
that increase upon each generation. Since that integer is a static variable,
we need to make sure to reset it upon each invocation of the `xdvipdfmx`
engine.