tectonic/xetex0.c: emit a Tectonic warning if missing characters are encountered
I've recently gotten a few bug reports from people exploring Tectonic's
Unicode capabilities that are not aware of the need to use the `fontspec`
package and to choose an appropriate font. Here we add a Tectonic warning
message that will hopefully point them in the right direction. In plain XeTeX,
the warning is only printed if \tracelostchars is activated, but this seems
like a big enough issue (and one that I don't think will happen spuriously?)
that I think it is reasonable to always print a message. I can revisit if
there are cases where the message gets annoying.

I hope that the help message is actually helpful for non-experts. If not,
perhaps we should create a dedicated forum thread and point people to it.

Closes #173.