src/io: add the Bundle trait
This lets us identify I/O backends that are "bundles" from the user
perspective -- large collections of TeX files. From the software architecture
perspective, these are IoProviders that can be used to generate format files
in a fully standalone fashion.

This trait will be used to improve the format caching code such that the
runtime options to specify a new Zip bundle or web-based bundle will actually
work again.