Fix format file caching, and therefore the -b and -w options.
The -b and -w options have been broken for a long time due to the switch to
on-the-fly creation of format files: I added a mechanism to cache format
files, but it was only active when using the default bundle. In this commit,
we split out the format file cache to be active whenever any kind of bundle is
being used.

The main work is to split out the part of
`tectonic::io::local_cache::LocalCache` that cached format files into a highly
similar type, `tectonic::io::format_cache::FormatCache`. Then there's some
plumbing to be done to activate the cache no matter how the active bundle is
specified.