tectonic: clean up `xeq_level`
This macro was being used in a gross way to hide an offset in the array
indexing. Make it a bit clearer what's going on, and importantly change some
magic hardcoded numbers into symbolic constants. Also fix an off-by-one in the
size that was exposed in the port to Winows.