tectonic/dpx-dpxfile.c: rip out some cache code for Windows
This is more code that should probably disappear anyway -- I think it only
exists for shelling out to Ghostscript and other stuff along those lines that
we definitely don't want to do in Tectonic. It used <dirent.h> and so was
going to be annoying to compile on Windows, so, sayonara.