tectonic: get rid of the zmem/mem distinction
I just have trouble believing that this makes a performance difference with a
modern compiler. I had been wary of doing this kind of transform before
because I did it with the eqtb variable earlier and thought that I *did* see a
big performance hit, but I now believe that was due to an unrelated change.