tests: centralize format-file generation
The `executable` test was failing intermittently on Windows. This seems to be
because of a bug in the executable test rig, where the program was being
passed a `--format` argument, but it was using a relative path that didn't
work when the test program was being run from a temporary directory. As such,
the test programs were all generating their own plain.fmt files, and managing
to stomp on each other's toes.

One solution is to be better about centralizing the code that generates the
plain.fmt file used for testing. This way, the executable tests avoid needing
to touch the network.