Deployment to AUR worked!
After making the `tectonic-deploy` use a co-maintainer, automagical deployment
to AUR worked. Go back to only running the deploy script when new tags are
pushed to master, and add a note in the PKGBUILD file since I could imagine
people trying to update the version on aur.archlinux.org without realizing
that it will get overwritten when new releases are made.