tectonic: defuse linebreaking in semantic pagination mode!
After much, much time spent trying to understand the linebreaking algorithm, I
think I am confident in the right way to modify it to activate semantic
pagination mode. Here it is! A quick test on xenia has visual results exactly
as I was hoping: a very "short" document with normal-looking paragraph lines
flying off the edge of the page. Huzzah!