tectonic/xetex-pagebuilder.c: customizations for semantic pagination mode
I am pretty sure this patch implements the semantic changes I would like to
have in order for semantic pagination to work the way I'd like. However,
the xenia test doesn't compile when I test it out:

```
[1
! Undefined control sequence.
\contentsline@latex ...@safe@activestrue \edef \x
                                                  {\endgroup \def \noexpand ...
l.175

Output written on paper.xdv (13 pages, 3904 bytes).
```

That's a problem! I infer that maybe the LaTeX output routine does stuff that
can't just be skipped. That will really throw a monkey wrench in things. To be
investigated.

(I can get farther in the processing if I comment out the figures in the test
file, so the main problem seems to be related to float processing ... which is
a bit promising since floats need to be hacked to work anyway. But I also got
an error at the end of the file, perhaps at a "regular" pagebreak given that
in S.P.-mode the paragraphs are all one line long.)