src/lib.rs: add an all-in-one function `latex_to_pdf()`
This gives us a nice concise example for the very top of the Tectonic API
documentation. And hopefully it is actually useful for people!

The testing infrastructure got a big revamp to allow us to actually execute
the doctest on the new API docs. Doctest and integration-test executables are
linked against the standard build of the main crate, so the supporting test
infrastructure has to live in that crate and can't be gated behind a
`#[cfg(test)]` attribute. I figured out a way to make this happen that I'm
happy with — it makes these tests possible, but should only bloat the
crate by a small amount.

Importantly, this approach means that we can also make it so that the tests in
`tests/executable.rs`, which launch the build `tectonic` binary, leverage the
same test mode. This is a nice win — before, these tests would go to the
network and pull down assets into the local user's cache. Now they use the
same set of test assets as everybody else.