Suppress unknown warnings on Clang
For this line:

    #pragma GCC diagnostic ignored "-Wstringop-overflow"

Clang prints this warning:

    warning: unknown warning group '-Wstringop-overflow', ignored [-Wunknown-warning-option]

This change imports a library with clever CPP that suppresses warnings
for GCC and avoids showing warnings for Clang.

I imported the whole library to include the license and helpful
documentation.

I renamed the header files to hint that push and pop need to be included
in equal (nested) numbers.