dist: fix the static Linux build (?)
After much wailing and gnashing of teeth, I think I have an approach that
fixes the static Linux build in the face of proc-macro usage. In particular,
proc-macros can work if we cross-compile. Here, we set up a Docker container
with a fakey cross-compilation setup that farms out the work to an Alpine
Linux chroot. That way we can take advantage of the statically linked,
MUSL-based libraries that Alpine provides.

We can now build static Tectonic without having to gate any features and using
Rust stable, so this should update the CI to not ignore the problem if the
static build fails.