.travis.yml: disable caching
The Travis docs say that caching is probably not a net win if you're caching
more than a few hundred megabytes of deps, and we are. Also, I recall seeing
other Rust projects disable Travis caching for performance reasons. I recently
had to clear the caches because they were getting so big that Travis was
timing out.